Deploy/Frontend/requirements.txt
streamlit
pandas
numpy
scikit-learn==1.0.2
requests
tensorflow-cpu
typing_extensions